Fix create news admin
