+++ app/Http/ViewComposers/IndexComposer.php
@@ -1,0 +1,20 @@
+<?php
+namespace App\Http\ViewComposers;
+
+use Illuminate\View\View;
+use App\Models\Category;
+
+class IndexComposer
+{
+    /**
+     * Bind data to the view.
+     *
+     * @param View $view of view
+     *
+     * @return void
+     */
+    public function compose(View $view)
+    {
+        $view->with('categories', Category::all());
+    }
+}
