+++ config/constant.php
@@ -1,0 +1,4 @@
+<?php
+    return [
+        'path_upload' => 'upload/picture_news',
+    ];
