+++ public/admin/js/editor.js
@@ -1,3 +1,3 @@
 $(function () {
-    CKEDITOR.replace('editor1');
-  });
+    $(".textarea").wysihtml5();
+    });
